film_backend/requirements.txt
flask
Flask-SQLAlchemy
psycopg2-binary
sqlalchemy
flask_sqlalchemy
flask_cors
PyJWT